requirements.txt
Flask
firebase-admin
pyrebase4